requirements.txt
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
dnspython==1.16.0
entrypoints==0.3
flake8==3.7.9
future==0.18.2
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pymongo==3.10.1
PySocks==1.7.1
python-telegram-bot==12.2.0
six==1.13.0
SQLAlchemy==1.3.12
tornado==6.0.3
